% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-points.R
\name{acnk}
\alias{acnk}
\title{Returns the values for the A/CNK = 1 plane in the appropriate coord sys}
\usage{
acnk(coordSys = "gc")
}
\arguments{
\item{coordsys}{coordinate system to use.
gc: granite coordinates (Si,Al,Ca,NK = Na+K,FM = Fe+Mg)}
}
\description{
Returns the values for the A/CNK = 1 plane in the appropriate coord sys
}
