% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-points.R
\name{cn.gc}
\alias{cn.gc}
\title{Granite coordinates
Granite coords are defined by (Si,Al,Ca,NK = Na+K,FM = Fe+Mg)}
\usage{
cn.gc()
}
\description{
Granite coordinates
Granite coords are defined by (Si,Al,Ca,NK = Na+K,FM = Fe+Mg)
}
