% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drawggTetraPlot}
\alias{drawggTetraPlot}
\title{Draw an empty canvas with the edges of tetraedron}
\usage{
drawggTetraPlot(
  apicesNames = c("A", "B", "C", "D"),
  axes = F,
  zscale = 3.5,
  show = T
)
}
\arguments{
\item{apicesNames}{Names of the apices to be drawn on plot}

\item{axes}{(F) Boolean, plot the cartesian axes?}

\item{zscale}{A magical number to be passed to rgldev, adjusting the vertical scale.
Apparently 3.5 looks nice and gives apparent square aspect ratio.}

\item{show}{(T) Boolean. If true, immediately show the plot otherwise just return it.}
}
\description{
Draw an empty canvas with the edges of tetraedron
}
