% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepareDataMat}
\alias{prepareDataMat}
\title{Convenience function. Assemble and converts a matrix from a vector}
\usage{
prepareDataMat(data, coord_names, object_names, object_types = "Nature")
}
\arguments{
\item{data}{a vector that holds the data}

\item{coord_names}{Names of the coordinates (will be colnames)}

\item{object_names}{Names of the objects (will be rownames)}

\item{object_types}{String describing what the "objects" are (minerals, end-members...)
Will be used in the tibble to identify them}
}
\description{
Convenience function. Assemble and converts a matrix from a vector
}
