% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tetraCoords}
\alias{tetraCoords}
\title{Calculate the (x,y,z) plotting coordinates from the values of 4 components}
\usage{
tetraCoords(A, B, C, D)
}
\arguments{
\item{{A, B, C, D}}{Coordinates of a point in (a,b,c,d) form}
}
\description{
Calculate the (x,y,z) plotting coordinates from the values of 4 components
}
